# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/trips": {
      "post": {
        "description": "One day per date, or one \"Day 1\" without dates.",
        "operationId": "Trips_create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "A client-generated UUID, new per action and reused on its retries. For 24 hours a retry with the same key and body returns the first response (header `Idempotent-Replayed: true`); the same key with a different request answers 422 IDEMPOTENCY_KEY_REUSED, and while the first is still running 409 IDEMPOTENCY_KEY_IN_PROGRESS.",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTripDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a trip",
        "tags": [
          "trips"
        ]
      }
    },
    "/v1/trips/{id}": {
      "get": {
        "operationId": "Trips_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Trip with members, days and markers",
        "tags": [
          "trips"
        ]
      },
      "patch": {
        "operationId": "Trips_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTripDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Owner: title, dates, visibility",
        "tags": [
          "trips"
        ]
      },
      "delete": {
        "operationId": "Trips_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Owner: delete the trip with everything in it",
        "tags": [
          "trips"
        ]
      }
    },
    "/v1/trips/{id}/copy": {
      "post": {
        "description": "Copies title, dates, days and markers into a new trip I own (my default visibility). No photos, comments, likes or members. Own and private trips: 403 TRIP_NOT_COPYABLE.",
        "operationId": "Trips_copy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "A client-generated UUID, new per action and reused on its retries. For 24 hours a retry with the same key and body returns the first response (header `Idempotent-Replayed: true`); the same key with a different request answers 422 IDEMPOTENCY_KEY_REUSED, and while the first is still running 409 IDEMPOTENCY_KEY_IN_PROGRESS.",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a public trip to my trips",
        "tags": [
          "trips"
        ]
      }
    },
    "/v1/trips/{id}/days": {
      "post": {
        "description": "Extends endDate by one day for dated trips.",
        "operationId": "Trips_addDay",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "A client-generated UUID, new per action and reused on its retries. For 24 hours a retry with the same key and body returns the first response (header `Idempotent-Replayed: true`); the same key with a different request answers 422 IDEMPOTENCY_KEY_REUSED, and while the first is still running 409 IDEMPOTENCY_KEY_IN_PROGRESS.",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripDayDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Append a day",
        "tags": [
          "trips"
        ]
      }
    },
    "/v1/days/{id}": {
      "delete": {
        "description": "Later days move up. The last remaining day can’t be deleted (fields.id = [\"lastDay\"]).",
        "operationId": "Trips_deleteDay",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a day and its markers",
        "tags": [
          "trips"
        ]
      }
    },
    "/v1/trips/{id}/members": {
      "post": {
        "operationId": "Trips_addMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMemberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripMemberDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Owner adds an editor by username. Idempotent.",
        "tags": [
          "trips"
        ]
      }
    },
    "/v1/trips/{id}/members/{userId}": {
      "delete": {
        "operationId": "Trips_removeMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Owner removes an editor, or an editor leaves (own userId)",
        "tags": [
          "trips"
        ]
      }
    },
    "/v1/me/trips": {
      "get": {
        "operationId": "Trips_myTrips",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "nextCursor from the previous page",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripSummaryPageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Trips I own or collaborate on, most recently changed first",
        "tags": [
          "trips"
        ]
      }
    },
    "/v1/me/stats": {
      "get": {
        "operationId": "Trips_stats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeStatsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "My trip, marker and photo counts",
        "tags": [
          "trips"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tripinly API",
    "description": "REST API for the Tripinly apps. Errors use `{ error: { code, message, details } }`; switch on `code`.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UserSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string",
            "example": "jonas.k"
          },
          "displayName": {
            "type": "string",
            "example": "Jonas"
          }
        },
        "required": [
          "id",
          "username",
          "displayName"
        ]
      },
...
      "TripSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "nullable": true
          },
          "endDate": {
            "type": "string",
            "nullable": true
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "private"
            ]
          },
          "owner": {
            "$ref": "#/components/schemas/UserSummaryDto"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "editor",
              "viewer"
            ]
          },
          "coverThumbUrl": {
            "type": "string",
            "nullable": true,
            "description": "First cover photo in the trip (from stage 5)"
          },
          "dayCount": {
            "type": "number"
          },
          "markerCount": {
            "type": "number"
          },
          "likeCount": {
            "type": "number"
          },
          "likedByMe": {
            "type": "boolean"
          },
          "copyCount": {
            "type": "number"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "startDate",
          "endDate",
          "visibility",
          "owner",
          "role",
          "coverThumbUrl",
          "dayCount",
          "markerCount",
          "likeCount",
          "likedByMe",
          "copyCount",
          "updatedAt"
        ]
      },
...
      "CreateTripDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "Vienna weekend"
          },
          "startDate": {
            "type": "string",
            "example": "2026-10-09",
            "description": "Without dates the trip has one \"Day 1\""
          },
          "endDate": {
            "type": "string",
            "example": "2026-10-11",
            "description": "Needs startDate; one day is created per date (max 20)"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "private"
            ],
            "description": "Defaults to the user's defaultTripVisibility"
          },
          "memberUsernames": {
            "maxItems": 20,
            "description": "Added as editors",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title"
        ]
      },
      "TripMemberDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserSummaryDto"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "editor"
            ]
          }
        },
        "required": [
          "user",
          "role"
        ]
      },
      "CopiedFromDto": {
        "type": "object",
        "properties": {
          "tripId": {
            "type": "string",
            "format": "uuid"
          },
          "owner": {
            "$ref": "#/components/schemas/UserSummaryDto"
          }
        },
        "required": [
          "tripId",
          "owner"
        ]
      },
      "LocationDto": {
        "type": "object",
        "properties": {
          "lat": {
            "type": "number",
            "example": 48.2082,
            "minimum": -90,
            "maximum": 90
          },
          "lng": {
            "type": "number",
            "example": 16.3738,
            "minimum": -180,
            "maximum": 180
          }
        },
        "required": [
          "lat",
          "lng"
        ]
      },
      "MarkerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tripId": {
            "type": "string",
            "format": "uuid"
          },
          "dayId": {
            "type": "string",
            "format": "uuid"
          },
          "placeId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "location": {
            "$ref": "#/components/schemas/LocationDto"
          },
          "time": {
            "type": "string",
            "nullable": true,
            "example": "14:30"
          },
          "position": {
            "type": "number"
          },
          "coverPhotoId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "coverThumbUrl": {
            "type": "string",
            "nullable": true,
            "description": "Signed URL, expires after ~1 h"
          },
          "photoCount": {
            "type": "number"
          },
          "likeCount": {
            "type": "number"
          },
          "likedByMe": {
            "type": "boolean"
          },
          "commentCount": {
            "type": "number"
          },
          "createdBy": {
            "nullable": true,
            "description": "Null after the creator deleted their account",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserSummaryDto"
              }
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "tripId",
          "dayId",
          "placeId",
          "name",
          "location",
          "time",
          "position",
          "coverPhotoId",
          "coverThumbUrl",
          "photoCount",
          "likeCount",
          "likedByMe",
          "commentCount",
          "createdBy",
          "createdAt",
          "updatedAt"
        ]
      },
      "TripDayDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "position": {
            "type": "number",
            "description": "0-based; the client shows \"Day n\" as position + 1"
          },
          "date": {
            "type": "string",
            "nullable": true,
            "example": "2026-10-09",
            "description": "startDate + position, or null"
          },
          "markers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MarkerDto"
            }
          }
        },
        "required": [
          "id",
          "position",
          "date",
          "markers"
        ]
      },
      "TripDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "nullable": true,
            "example": "2026-10-09"
          },
          "endDate": {
            "type": "string",
            "nullable": true,
            "example": "2026-10-11"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "private"
            ]
          },
          "owner": {
            "$ref": "#/components/schemas/UserSummaryDto"
          },
          "myRole": {
            "type": "string",
            "enum": [
              "owner",
              "editor",
              "viewer"
            ]
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TripMemberDto"
            }
          },
          "likeCount": {
            "type": "number"
          },
          "likedByMe": {
            "type": "boolean"
          },
          "copyCount": {
            "type": "number"
          },
          "copiedFrom": {
            "nullable": true,
            "description": "Shown as \"Copied from @handle\" while the source exists",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CopiedFromDto"
              }
            ]
          },
          "days": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TripDayDto"
            }
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "startDate",
          "endDate",
          "visibility",
          "owner",
          "myRole",
          "members",
          "likeCount",
          "likedByMe",
          "copyCount",
          "copiedFrom",
          "days",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateTripDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "startDate": {
            "type": "string",
            "nullable": true,
            "example": "2026-10-09",
            "description": "null removes the dates (days stay, shown as \"Day n\")"
          },
          "endDate": {
            "type": "string",
            "example": "2026-10-11",
            "description": "Resizes the trip: adds empty days, or removes trailing days if they have no markers (else 400 fields.endDate = [\"daysNotEmpty\"])"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "private"
            ]
          }
        }
      },
      "AddMemberDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "anna.n"
          }
        },
        "required": [
          "username"
        ]
      },
      "TripSummaryPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TripSummaryDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
      "MeStatsDto": {
        "type": "object",
        "properties": {
          "tripCount": {
            "type": "number",
            "description": "Trips I own or collaborate on"
          },
          "markerCount": {
            "type": "number",
            "description": "Markers I created"
          },
          "photoCount": {
            "type": "number",
            "description": "Photos I uploaded (from stage 5)"
          }
        },
        "required": [
          "tripCount",
          "markerCount",
          "photoCount"
        ]
      },
...
    }
  }
}